backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "langchain>=0.2.5",
    "langchain-community>=0.1.0",
    "langgraph>=1.1.3",
    "langsmith>=0.1.0",
    "pydantic[email]>=2.12.5",
    "httpx>=0.24.0",
    "cryptography>=42.0.0",
    "bcrypt>=4.1.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "tiktoken>=0.6.0",
    "uvicorn>=0.27.0",
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
demo = [
    "streamlit>=1.28.0",
]
